+++ versions/26.1.x/tests/LinearTimingInstrumentationTest.java
@@ -1,4 +1,4 @@
-package net.sexidium;
+package net.linear;
 
 import static org.junit.jupiter.api.Assertions.*;